// src/definitions/enums.tsx
export enum TileKeys {
  utilityOverview = "utilityOverview",
  calculationsPerUtility = "calculationsPerUtility",
  RTOPerUtility = "RTOPerUtility",
  eBoks = "eBoks"
}

export enum TileActions {
  move = "move",
  resize = "resize"
}

export enum TileGroup {
  heatIntelligence = "Heat intelligence",
  returnTemperatureOptimizer = "Return temperature optimizer",
  eBoks = "e-Boks"
}